fix: resolve build errors (missing includes, format warning, enum cast, buffer size)
- Add missing #include "esp_timer.h" to wifi_manager.cpp and discovery.cpp
- Cast total to int in main.cpp to fix format string type mismatch
- Cast enum bitwise OR to lv_style_selector_t in ui.cpp to fix -Werror
- Increase snprintf buffer from 16 to 24 in ui.cpp to fix truncation warning

+++ main/main.cpp
@@ -63,7 +63,7 @@ static void on_timer_started(void *, esp_event_base_t, int32_t, void *data) {
   auto total = *static_cast<int32_t *>(data);
   char label[64];
   timer_get_label(label, sizeof(label));
-  ESP_LOGI(TAG, "Timer started: %ds '%s'", total, label);
+  ESP_LOGI(TAG, "Timer started: %ds '%s'", (int)total, label);
   ui_timer_show(total, label);
 }
